requirements.txt
absl-py==0.8.0
aiohttp==3.6.2
aioredis==1.3.1
apache-beam==2.25.0
appdirs==1.4.3
argon2-cffi==20.1.0
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
avro-python3==1.9.1
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
blessings==1.7
blinker==1.4
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
Click==7.0
cloud-init==20.1
colorama==0.4.3
colorful==0.5.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crcmod==1.7
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
easydict==1.9
entrypoints==0.3
fairness-indicators==0.25.0
fastavro==1.2.0
filelock==3.0.12
future==0.18.2
gast==0.3.3
google==3.0.0
google-api-core==1.22.2
google-api-python-client==1.12.8
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.29.0
h5py==2.10.0
hdfs==2.5.8
hiredis==1.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==3.3.0
incremental==16.10.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jax==0.1.70
jaxlib==0.1.47
jedi==0.17.2
Jinja2==2.10.1
joblib==0.14.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
macaroonbakery==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
more-itertools==4.2.0
msgpack==1.0.0
multidict==4.7.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
notebook==6.1.5
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.0
opencensus==0.7.10
opencensus-context==0.1.1
opencv-python==4.4.0.46
opt-einsum==3.2.1
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.3
parso==0.7.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.12.2
psutil==5.7.2
ptyprocess==0.6.0
py-spy==0.3.3
pyarrow==0.17.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pydot==1.4.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.11.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytorch-lightning==0.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==17.0.0
qtconsole==5.0.1
QtPy==1.9.0
ray==0.8.7
redis==3.4.1
requests==2.25.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
screen-resolution-extra==0.0.0
seaborn==0.11.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
soupsieve==2.0.1
ssh-import-id==5.10
systemd-python==234
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.1
tensorflow==2.3.1
tensorflow-data-validation==0.25.0
tensorflow-datasets==4.1.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.25.0
tensorflow-model-analysis==0.25.0
tensorflow-serving-api==2.3.0
tensorflow-transform==0.25.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tfco-nightly==0.3.dev20201124
tfx-bsl==0.25.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.6.0
torchsummary==1.5.1
torchvision==0.7.0
tornado==6.1
tqdm==4.50.2
traitlets==5.0.5
Twisted==18.9.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
witwidget==1.7.0
wrapt==1.12.1
xkit==0.0.0
yarl==1.5.1
zipp==1.0.0
zope.interface==4.7.1
